feat: extract jobs data
