+++ src/data/jobs.json
@@ -1,0 +1,26 @@
+{
+  "resourcify": {
+    "href": "https://resourcify.com/",
+    "classes": "ms-1 inline",
+    "image": "https://res.cloudinary.com/dwrurydlt/image/upload/v1693067033/resourcify_69f3b5b70d.webp",
+    "alt": "Resourcify",
+    "width": 106,
+    "height": 21
+  },
+  "bjss": {
+    "href": "https://bjss.com/",
+    "classes": "mx-1 inline",
+    "image": "https://res.cloudinary.com/dwrurydlt/image/upload/v1692645367/bjss_180dc7fdd7.webp",
+    "alt": "BJSS",
+    "width": 45,
+    "height": 25
+  },
+  "booking": {
+    "href": "https://booking.com/",
+    "classes": "ms-1 inline",
+    "image": "https://res.cloudinary.com/dwrurydlt/image/upload/v1693067075/bookingcom_91b7aa2e36.svg",
+    "alt": "Booking.com",
+    "width": 115,
+    "height": 25
+  }
+}
